requirements.txt
asgiref==3.7.2
beautifulsoup4==4.12.3
bs4==0.0.2
cachetools==5.3.2
certifi==2023.11.17
charset-normalizer==3.3.2
coreapi==2.3.3
coreschema==0.0.4
Django==5.0
django-cors-headers==4.3.1
django-elasticsearch-dsl==8.0
django-rest-swagger==2.2.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7
elastic-transport==8.11.0
elasticsearch==8.11.1
elasticsearch-dsl==8.11.0
google-api-core==2.16.1
google-api-python-client==2.116.0
google-auth==2.27.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
googleapis-common-protos==1.62.0
grobid-tei-xml==0.1.3
httplib2==0.22.0
idna==3.6
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.3
lxml==5.1.0
Markdown==3.5.1
MarkupSafe==2.1.5
mysqlclient==2.2.1
oauthlib==3.2.2
openapi-codec==1.3.2
packaging==23.2
protobuf==4.25.2
psycopg2-binary==2.9.9
pyasn1==0.5.1
pyasn1-modules==0.3.0
PyJWT==2.8.0
pyparsing==3.1.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
PyYAML==6.0.1
requests==2.31.0
requests-oauthlib==1.3.1
rest-framework-simplejwt==0.0.2
rsa==4.9
simplejson==3.19.2
six==1.16.0
soupsieve==2.5
sqlparse==0.4.4
typing_extensions==4.9.0
tzdata==2023.3
uritemplate==4.1.1
urllib3==2.1.0
